Fix shouldPreserveAddressOrderOnInsertions test
The test failed because the ArrayList - in contrast to the underlying Set - did not check for duplicates on insert. Under certain circumstances this lead to an index out
of bounds exception because the list in the test contained duplicated entries which were not present in the set of the Socks5Proxy.
I fixed the issue by only inserting the address when it was not in the list before.